jRafferty_Music: Upload 1
I constructed a test within the html file to figure out whether my css file successfully linked to my converted webfonts.
As of this upload, this test has not achieved a modicum of success.
Troubleshooting is required. More detailed descriptions of my issues are commented directly into both the css and html files.