Upload 5
What I Did: I built the tour page. I solved all of the problems I faced as I found them, and the execution is nearly flawless. I also modified the hover states for the nav to expand both above and below the nav bar, effectively making the effect bidirectional.

What I Need Help With: I still haven't figured out why the nav bar is stuck and won't travel with the viewport, but I'm sure it's a simple fix. I'm also interested in adding a jump tag into my html so that the nav items autoscroll past the home page and land on the unique page contents.