Upload 6
What I Did: Built the basic structure of the store page. I still need to work on the cart function before the store page is complete and can be adapted to tablet and mobile formats, but I plan on doing as much media query work with the other pages, which are feasibly complete in desktop mode.

What I Need Help With: The one glaring issue I see with the current design, aside from the current absence of an interactive cart, is the fact that the bulletin page is static. I need your help in laying out the structural foundation of the bulletin page so that it can change to new content every few seconds as specified by a timer function, which I assume has something to do with some sort of 'for' loop function in mainJS. I remember fireworks going off in my brain when we were discussing your idea to make it work, but I have since lost that idea somewhere in the mire of undercooked scrambled eggs that my brain as since become.