Upload 8: The Penultimate Commit
What I Did: I have almost entirely completed this website. The store page is fully constructed and fleshed out, there shouldn't be a placeholder in sight, and media query styles are all tweaked and finalized. The tour pages for all three device sizes are nearly pixel-perfect, and I proofread and bug-fixed the bulletin text one final time.

What I Need Help With: The mobile size media query's off-canvas nav obscures my shopping cart indicator entirely. This manifests no new bugs with the mobile interaction, other than the absence of any kind of indicator of how many items are in the cart. Since the off-canvas nav was a part of your edits and involves some amount of j-Query, I lack the ability to parse its inner workings or make the necessary edits to it. Ideally, the cart would be relocated to the nav bar on mobile, opposite from the hamburger icon.

What I Have To Say: Thank you for equipping me with the skills necessary to simultaneously create and solve as many problems as I have during this course. Outside of coding, user interface design, and all of the new tools that you have taught me, your course has made me a more creative and capable person in general. I will forever be indebted to you as I proceed in my professional endeavors, and I hope to keep in touch long after this week is over. Thank you.