Add files via upload
I got the cart to show up on desktop and mobile, however there is some trickier JS to have multiple add to carts connect to the same cart counter. So it works on desktop, mobile though it won't, which is honestly OK...you've definitely go over and beyond on this project.